packages/relaymd-worker/pyproject.toml
[project]
name = "relaymd-worker"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
    "relaymd-core",
    "relaymd-api-client",
    "infisical-python>=2.0,<2.2",
    "httpx[socks]",
    "pydantic-settings>=2.0",
    "nvidia-ml-py",
    "orjson",
    "structlog",
    "tenacity>=8.2",
]

[project.scripts]
relaymd-worker = "relaymd.worker.__main__:main"

[tool.uv.sources]
relaymd-core = { workspace = true }
relaymd-api-client = { workspace = true }

[tool.pytest.ini_options]
pythonpath = ["src"]
